% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jtrans.R
\name{jtrans}
\alias{jtrans}
\title{Johnson Transformation for Normality}
\usage{
jtrans(x, test = "shapiro.test", exclude.original = TRUE, z.length = 101)
}
\arguments{
\item{x}{the non-normal numerical data.}

\item{test}{the normality test used to select fits.}

\item{exclude.original}{whether the original data should be excluded when comparing fits.}

\item{z.length}{the length of the z vector. It's also the number of different fits estiamted in the algorithm. Set larger z.length value if you want extra precision.}
}
\value{
A list with class "jtrans" containing the following components:
 \item{original}{original data.}
 \item{transformed}{transformed data.}
 \item{type}{type of transformation selected.}
 \item{test}{normality test used to select transformations.}
 \item{z}{selected z value among 101 values from 0.25 to 1.25.}
 \item{eta, gamma, lambda, epsilon}{transformation parameters.}
 \item{p.value}{the maximum p.value returned by test}
}
\description{
\code{jtrans} transforms any continuous univariate vector to a vector from standard normal distribution.
}
\details{
\code{jtrans} fits data to a set of distributions from Johnson family. A normality test is used (Shapiro-Wilk test by default) to find the best fit by choosing the fit with maximum p.value under that given test. It returns the transformed data, the corresponding type of Johnson curve and parameter estimations.

Since the default Shapiro-Wilk test can only accept sample size between 3 and 5000, there will be a problem when sample size is larger than 5000. In such cases, one should specify another normality test in the test parameter, generally the \code{ad.test} in the \pkg{nortest} package is recommended.

Sometimes, this algorithm may return poor fits. The most extreme case is that all the transformed data have smaller p.values than the original data's. In such cases, the \code{exclude.original} flag should be set to FALSE, so \code{jtrans} will return the original data as the transformed data.
}
\examples{
# generate 100 non-normal data and transform it.
x <- rexp(100, .2)
jt <- jtrans(x)
jt
}
\references{
Chou, Y. M., Polansky, A. M., & Mason, R. L. (1998).
Transforming non-normal data to normality in statistical process control.
Journal of Quality Technology, 30(2), 133-141.
}

