% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generics.R
\name{trans_func}
\alias{trans_func}
\title{Export the transformation into LaTeX formula}
\usage{
trans_func(y, latex = TRUE)
}
\arguments{
\item{y}{an jtrans object}

\item{latex}{whether to print the latex formula of the transformation}
}
\description{
\code{trans_func} turns a \code{jtrans} object into either an R function of this transformation for future use or a LaTeX formula for display in the report.
}

